tst-core: ST 0903 VTargetPack inner walk to BER-OID (validate-1 E5 follow-up)
The top-level ST 0903 LS walker migrated to BER-OID in plan E5
(031b3c4), but the nested VTargetPack body walk in
vtarget_pack::decode::read_pack still read tag IDs as `cursor[0]`
(raw single byte). Current ST 0903.6 §10.2 pack tags top out at 107
so no present bug, but a future ST 0903.7+ pack tag >= 128 would
break framing identically to the pre-E5 top-level bug — the first
BER-OID continuation byte (0x81 0x..) would be read as the start of
the BER length and misframe the entire walk.

Migrate the inner walk to BER-OID per ST 0107.5 §6.3.1, matching
the post-E5 top-level walker pattern. Forward-compat tags >= 128
land in `pack.unknown` keyed by their decoded u32 tag value.

Encode side: replace the silent-drop guard (`if field.tag <= 0xFF`)
with write_ber_oid + write_ber emission, and update encoded_len to
mirror via ber_oid_len(field.tag) + ber_len(value.len()) + value.len().

VTargetPackError tag fields widen u8 -> u32 to faithfully represent
multi-byte BER-OID tag IDs in error paths. Matches KlvFieldError
which already uses u32 throughout. Public-api baseline updated; no
new #[non_exhaustive] enums; BASELINE stays at 134.

Byte-identical preservation for defined tags 1..=107: BER-OID of any
value <= 127 is a single byte equal to the value, so legacy wire is
unchanged. A new test (defined_pack_tags_byte_identical_pre_and_post_e5_followup)
pins a hand-built byte sequence; unknown_tag_multibyte_ber_oid_round_trips
covers the 127/128/16383/16384 BER-OID size boundaries. The existing
unknown_tags_preserved test was constructed against the pre-fix
single-byte tag mental model and is updated to use proper BER-OID
encoding for tag 200 (0x81 0x48) on the wire.

vtarget_pack has no decode_strict variant — the strict path
(decode_vtarget_series_strict) reuses the same lenient read_pack
parser, so only the lenient walker needed migration. Closes E5's
"share substrate" intent at the inner-walk layer.

+++ crates/tst-core/src/klv/st0903/vtarget_pack/model.rs
@@ -250,21 +250,24 @@ pub(crate) fn pack_lookup(tag: u8) -> Option<&'static PackTagSpec> {
 pub enum VTargetPackError {
     #[error("truncated BER-OID Target ID")]
     TruncatedTargetId,
+    /// Field-level framing or value truncation. `tag` is BER-OID-
+    /// decoded (`u32`) to cover both the §10.2 typed universe (1..=107)
+    /// and forward-compat ST 0903.7+ multi-byte BER-OID tag IDs.
     #[error("VTargetPack tag {tag}: truncated value")]
-    TruncatedField { tag: u8 },
+    TruncatedField { tag: u32 },
     #[error("VTargetPack tag {tag}: declared length {declared} exceeds available {available}")]
     LengthOverrun {
-        tag: u8,
+        tag: u32,
         declared: usize,
         available: usize,
     },
     #[error("VTargetPack tag {tag}: malformed IMAPB value")]
-    MalformedImapb { tag: u8 },
+    MalformedImapb { tag: u32 },
     #[error("VTargetPack tag {tag}: malformed UTF-8 string")]
-    MalformedUtf8 { tag: u8 },
+    MalformedUtf8 { tag: u32 },
     #[error("VTargetPack tag {tag}: invalid value length {got} (expected {expected})")]
     InvalidLength {
-        tag: u8,
+        tag: u32,
         expected: usize,
         got: usize,
     },
